Add UTs for SubNav
